enough memory optimizations have been applied, next step is getting the algorithm working for facial recognition.

+++ src/main/java/com/github/retrooper/bigdata/KMeansClusterMain.java
@@ -12,8 +12,8 @@ import java.util.function.Supplier;
 
 public class KMeansClusterMain {
     public static void main(String[] args) {
-        double[] input = new double[]{
-                1.1, 1.1, 1.1, 1.4, 3, 3.2, 3.3, 3.4, 5, 5, 5
+        float[] input = new float[]{
+                1.1f, 1.1f, 1.1f, 1.4f, 3f, 3.2f, 3.3f, 3.4f, 5f, 5f, 5f
         };
 
         FunctionDataset2D function = new FunctionDataset2D(input);
@@ -26,7 +26,7 @@ public class KMeansClusterMain {
             System.out.println("What cluster should we put X in, good grades (1), mid grades (2), bad grades (3)");
             String line = scanner.nextLine();
             try {
-                double x = Double.parseDouble(line);
+                float x = (float) Double.parseDouble(line);
                 System.out.println("X: " + x + " in cluster: " + trainedModel.predict(new Point(x, 0)));
             } catch (Exception exception) {
                 break;
