+++ src/main/java/com/github/retrooper/bigdata/LinearRegressionMain.java
@@ -11,25 +11,25 @@ import java.util.function.Supplier;
 
 public class LinearRegressionMain {
     public static void main(String[] args) {
-        double[] input = new double[]{
-                -1, 0, 1, 2
+        float[] input = new float[]{
+                -1f, 0f, 1f, 2f
         };
 
-        double[] output = new double[]{
-                -2, 0, 2, 4
+        float[] output = new float[]{
+                -2f, 0f, 2f, 4f
         };
 
         FunctionDataset2D function = new FunctionDataset2D(input, output);
-        Supplier<LearningAlgorithm<Double>> dataSupplier = () -> LinearRegressionAlgorithm.build(function);
-        TrainingModel<Double> trainingModel = new TrainingModel<>();
-        ProductionModel<Double> trainedModel = trainingModel.train(dataSupplier);
+        Supplier<LearningAlgorithm<Float>> dataSupplier = () -> LinearRegressionAlgorithm.build(function);
+        TrainingModel<Float> trainingModel = new TrainingModel<>();
+        ProductionModel<Float> trainedModel = trainingModel.train(dataSupplier);
 
         Scanner scanner = new Scanner(System.in);
         while (true) {
             System.out.println("What X value should we predict based on the data?");
             String line = scanner.nextLine();
             try {
-                double x = Double.parseDouble(line);
+                float x = (float)Double.parseDouble(line);
                 System.out.println("X: " + x + ", y: " + trainedModel.predict(x));
             }
             catch (Exception exception) {
