+++ src/main/java/com/github/retrooper/bigdata/algorithm/LearningAlgorithm.java
@@ -1,5 +1,5 @@
 package com.github.retrooper.bigdata.algorithm;
 
 public interface LearningAlgorithm<T> {
-    double predict(T x);
+    float predict(T x);
 }
