+++ src/main/java/com/github/retrooper/bigdata/util/NDimensionalPoint.java
@@ -3,18 +3,18 @@ package com.github.retrooper.bigdata.util;
 import java.util.stream.Stream;
 
 public class NDimensionalPoint {
-    private final double[] coordinates;
+    private final float[] coordinates;
 
-    public NDimensionalPoint(double singularCoord) {
-        this(new double[]{singularCoord});
+    public NDimensionalPoint(float singularCoord) {
+        this(new float[]{singularCoord});
     }
 
-    public NDimensionalPoint(double[] coordinates) {
+    public NDimensionalPoint(float[] coordinates) {
         this.coordinates = coordinates;
     }
 
-    public NDimensionalPoint(Double[] coordinates) {
-        double[] array = new double[coordinates.length];
+    public NDimensionalPoint(Float[] coordinates) {
+        float[] array = new float[coordinates.length];
         for (int i = 0; i < array.length; i++) {
             array[i] = coordinates[i];
         }
@@ -22,7 +22,7 @@ public class NDimensionalPoint {
     }
 
     public double distanceSquared(NDimensionalPoint point) {
-        double distanceSquared = 0;
+        float distanceSquared = 0.0F;
         for (int i = 0; i < coordinates.length; i++) {
             double difference = coordinates[i] - point.coordinates[i];
             distanceSquared += difference * difference;
@@ -30,13 +30,13 @@ public class NDimensionalPoint {
         return distanceSquared;
     }
 
-    public double distance(NDimensionalPoint point) {
-        return Math.sqrt(distanceSquared(point));
+    public float distance(NDimensionalPoint point) {
+        return (float) Math.sqrt(distanceSquared(point));
     }
 
-    public double getCoordinatesSum() {
-        double sum = 0;
-        for (double c : coordinates) {
+    public float getCoordinatesSum() {
+        float sum = 0.0F;
+        for (float c : coordinates) {
             sum += c;
         }
         return sum;
@@ -47,7 +47,7 @@ public class NDimensionalPoint {
         return new NDimensionalPoint(coordinates());
     }
 
-    public double[] coordinates() {
+    public float[] coordinates() {
         return coordinates;
     }
