+++ src/main/java/com/github/retrooper/bigdata/util/Point.java
@@ -1,8 +1,8 @@
 package com.github.retrooper.bigdata.util;
 
 public class Point extends NDimensionalPoint {
-    public Point(double x, double y) {
-        super(new Double[] {x, y});
+    public Point(float x, float y) {
+        super(new Float[] {x, y});
     }
 
     @Override
@@ -10,11 +10,11 @@ public class Point extends NDimensionalPoint {
         return new Point(coordinates()[0], coordinates()[1]);
     }
 
-    public double x() {
+    public float x() {
         return coordinates()[0];
     }
 
-    public double y() {
+    public float y() {
         return coordinates()[1];
     }
 }
